refactor: move build_summary_item to widget utils
