+++ frappe/public/js/frappe/widgets/utils.js
@@ -95,6 +95,24 @@ function generate_grid(data) {
 	return grid_template_area
 }
 
+const build_summary_item = (summary) => {
+	let df = {fieldtype: summary.datatype};
+	let doc = null;
+
+	if (summary.datatype == "Currency") {
+		df.options = "currency";
+		doc = {currency: summary.currency};
+	}
+
+	let value = frappe.format(summary.value, df, null, doc);
+	let indicator = summary.indicator ? `indicator ${ summary.indicator.toLowerCase() }` : '';
+
+	return $(`<div class="summary-item">
+		<span class="summary-label small text-muted ${indicator}">${summary.label}</span>
+		<h1 class="summary-value">${ value }</h1>
+	</div>`);
+};
+
 // function get_luminosity(color) {
 // 	let c = color.substring(1);      // strip #
 // 	let rgb = parseInt(c, 16);   // convert rrggbb to decimal
@@ -127,4 +145,4 @@ function generate_grid(data) {
 //     return "#"+RR+GG+BB;
 // }
 
-export { generate_route, generate_grid };
+export { generate_route, generate_grid, build_summary_item };
